Scenes

+++ Engine/source/Handlers/GuiHandler.cpp
@@ -5,7 +5,7 @@ void GuiHandler::pushContainer(const std::shared_ptr<Container>& container)
 	this->containers.emplace_back(container);
 }
 
-void GuiHandler::updateGuis(const int& dt)
+void GuiHandler::updateGuis(const float& dt)
 {
 	for (auto& container : containers)
 		container->onUpdate(dt);
